+++ pkg/appsec/api_constraints_protection.go
@@ -12,8 +12,6 @@ type (
 	// The APIConstraintsProtection interface supports retrieving and updating API request constraints
 	// for a configuration and policy.
 	//
-	// Deprecated: this interface will be removed in a future release. Use the PolicyProtections interface instead.
-	//
 	// https://developer.akamai.com/api/cloud_security/application_security/v1.html#protections
 	APIConstraintsProtection interface {
 		// https://developer.akamai.com/api/cloud_security/application_security/v1.html#getapirequestconstraints